Extract time series of wrf file list of lat/lon

Read and extract data from a list of wrf output files and a table of lat/lon points based on the distance of the points and the center of model grid points, points outside the domain (and points on domain boundary) are not extracteds.

Usage

extract_serie(
  filelist,
  point,
  variable = "o3",
  field = "4d",
  prefix = "serie",
  new = "check",
  return.nearest = FALSE,
  fast = FALSE,
  use_ij = FALSE,
  latitude = "XLAT",
  longitude = "XLONG",
  use_TFLAG = FALSE,
  use_datesec = FALSE,
  id = "id",
  verbose = TRUE
)

Arguments

filelist

list of files to be read

point

data.frame with lat/lon

variable

variable name

field

'4d' (defoult), '3d', '2d' or '2dz' see notes

prefix

to output file, defolt is serie

new

TRUE, FALSE of 'check' see notes

return.nearest

return the data.frame of nearest points instead of extract the serie

fast

faster calculation of grid distances but less precise

use_ij

logical, use i and j from input instead of calculate

latitude

name of latitude coordinade variable in the netcdf

longitude

name of longitude coordinade variable in the netcdf

use_TFLAG

use the variable TFLAG (CMAQ / smoke) instead of Times (WRF)

use_datesec

use the variable date and datesec (WACCM / CAM-Chem) instead of Times (WRF)

id

name of the column with station names, if point is a SpatVector (points) from terra package

verbose

display additional information

Note

The field argument '4d' or '2dz' is used to read a 4d/3d variable droping the 3rd dimention (z).

new = TRUE create a new file, new = FALSE append the data in a old file, and new = 'check' check if the file exist and append if the file exist and create if the file doesnt exist

FOR CAMx time-series, use the options: use_TFLAG=T, latitude='latitude', longitude='longitude', new=T

FOR WACCM time-series, use the options: use_datesec=T, latitude='lat', longitude='lon', new=T

Examples

cat('Example 1: INMET stations for 2015\n')
#> Example 1: INMET stations for 2015
stations <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/BR-inmet_2015.Rds"))

cat('Example 2: METAR stations of Brazil\n')
#> Example 2: METAR stations of Brazil
stations <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/BR-metar.Rds"))

cat('Example 3: METAR soundings over Brazil\n')
#> Example 3: METAR soundings over Brazil
stations <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/BR-soundings.Rds"))

cat('Example 4: Global AERONET sites\n')
#> Example 4: Global AERONET sites
stations <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/Global-aeronet.Rds"))

cat('Example 5: Brazilian Air Quality: CETESB (SP), RAMQAr (ES) and SMAC (RJ)\n')
#> Example 5: Brazilian Air Quality: CETESB (SP), RAMQAr (ES) and SMAC (RJ)
stations <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/BR-AQ.Rds"))

files    <- dir(path = system.file("extdata",package="eva3dm"),
                pattern = 'wrf.day',
                full.names = TRUE)
dir.create(file.path(tempdir(),"SERIE"))
folder <- file.path(tempdir(),"SERIE")

# extract data for 3 locations
extract_serie(filelist = files, point = stations[1:3,],prefix = paste0(folder,'/serie'))
#> extracting series of o3 field 4d for 3 points
#> dim of lat/lon: 125 125 
#> used dim of lat/lon: 125 125 
#> calculating distances...
#>                 lat       lon source  i  j model_lat model_lon
#> Americana -22.72500 -47.34800 CETESB 82 37 -22.79819 -47.40500
#> Campinas  -22.94673 -47.11928 CETESB 83 37 -22.78904 -47.05811
#> Congonhas -23.61632 -46.66347 CETESB 84 34 -23.73756 -46.67905
#> reading o3 : /home/runner/work/_temp/Library/eva3dm/extdata/wrf.day1.o3.nc file 1 of 3 
#> reading o3 : /home/runner/work/_temp/Library/eva3dm/extdata/wrf.day2.o3.nc file 2 of 3 
#> reading o3 : /home/runner/work/_temp/Library/eva3dm/extdata/wrf.day3.o3.nc file 3 of 3 
#> output: /tmp/RtmprcJOnp/SERIE/serie.o3.Rds